Get AboutResource as account metadata for Drive V2.
In Drive V2 api, change ID is extended to int64, so we have to extend the size to 64bit.

For getting account metadata, WAPI:Drive V2 is not 1:1
mapping. Getting application list is separated from this
CL.

As refactoring gdata_wapi_feed_loader and making a separate
drive_api_feed_loader makes very redundant copy & paste,
so used IsDriveV2APIEnabled() for separating different JSON
feed handlings in completion callbacks.

CLs for getting file changelists and applists will follow.

BUG=chromium:127728
TEST=manually make sure this doesn't break FileBrowser


Review URL: https://chromiumcodereview.appspot.com/10823226

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@150718 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 6cdb24e2e4b49c1096bbd37693db3cc2eff822a6

+++ chrome/browser/chromeos/gdata/gdata_wapi_feed_processor.cc
@@ -49,14 +49,14 @@ GDataWapiFeedProcessor::~GDataWapiFeedProcessor() {
 
 GDataFileError GDataWapiFeedProcessor::ApplyFeeds(
     const std::vector<DocumentFeed*>& feed_list,
-    int start_changestamp,
-    int root_feed_changestamp,
+    int64 start_changestamp,
+    int64 root_feed_changestamp,
     std::set<FilePath>* changed_dirs) {
   bool is_delta_feed = start_changestamp != 0;
 
   directory_service_->set_origin(FROM_SERVER);
 
-  int delta_feed_changestamp = 0;
+  int64 delta_feed_changestamp = 0;
   FeedToFileResourceMapUmaStats uma_stats;
   FileResourceIdMap file_map;
   GDataFileError error = FeedToFileResourceMap(feed_list,
@@ -104,7 +104,7 @@ void GDataWapiFeedProcessor::UpdateFileCountUmaHistograms(
 
 void GDataWapiFeedProcessor::ApplyFeedFromFileUrlMap(
     bool is_delta_feed,
-    int feed_changestamp,
+    int64 feed_changestamp,
     FileResourceIdMap* file_map,
   std::set<FilePath>* changed_dirs) {
   DCHECK(BrowserThread::CurrentlyOn(BrowserThread::UI));
@@ -257,7 +257,7 @@ GDataDirectory* GDataWapiFeedProcessor::FindDirectoryForNewEntry(
 GDataFileError GDataWapiFeedProcessor::FeedToFileResourceMap(
     const std::vector<DocumentFeed*>& feed_list,
     FileResourceIdMap* file_map,
-    int* feed_changestamp,
+    int64* feed_changestamp,
     FeedToFileResourceMapUmaStats* uma_stats) {
   DCHECK(BrowserThread::CurrentlyOn(BrowserThread::UI));
   DCHECK(uma_stats);
